requirements.txt
retrying==1.3.3
beautifulsoup4==4.7.1
cssselect==0.9.1
feedfinder2==0.0.4
feedparser==5.2.1
idna==2.1
jieba3k==0.35.1
lxml==3.6.0
newspaper3k==0.2.8
nltk==3.2.1
Pillow==3.2.0
python-dateutil==2.5.3
PyYAML==3.11
requests==2.10.0
requests-file==1.4
six==1.10.0
tldextract==2.0.1
